Project/ingestion/requirements.txt
kaggle==1.6.17
pandas==2.2.3
pyarrow==18.1.0
google-cloud-storage==2.18.2
tqdm==4.67.1
click==8.1.8
